changing type of humidity field in WeatherSample to be an Integer. simplifying WeatherList

+++ backend/src/main/java/com/itay/weather/backend/dto/WeatherSample.java
@@ -16,5 +16,5 @@ public class WeatherSample {
     private Location location;
     private Timestamp time;
     private Double temperature;
-    private Double humidity;
+    private Integer humidity;
 }
