+++ miners/tomorrow-miner/src/main/java/com/itay/weather/tomorrowminer/dto/WeatherSample.java
@@ -17,5 +17,5 @@ public class WeatherSample {
     private Location location;
     private Timestamp time;
     private Double temperature;
-    private Double humidity;
+    private Integer humidity;
 }
